Address Copilot: PNG drift, chi_max typing, SA/SM nullspace semantics
All five findings acted on.

Both sync paths (docs-sync.yml, generate_docs.sh) copied the current PNGs but
never removed stale ones, so a renamed or deleted illustration lingered in
docs/source/notebooks indefinitely. Both now clean destination PNGs before
copying, mirroring the notebook clean-out.

chi_max is now annotated float on the five functions whose np.inf support this
PR added -- svd, qr, split_two_site_tensor, compress_bond, compress -- with
docstrings to match; typed callers could not use the documented idiom against
an int contract. Verified under both mypy 2.3.1 (pre-commit) and 1.15 (CI
lockfile). Functions that never claimed np.inf keep their int contract.

For SA/SM, a start vector the PSD effective density operator annihilates
exactly is itself an exact zero-eigenvalue eigenvector -- precisely what those
modes ask for -- so update_bond now uses it directly instead of the range
restart, which excludes the nullspace by construction. Empirically ARPACK's
internal random restarts happened to recover eigenvalue 0 through the range
path anyway, so this is determinism rather than a demonstrated wrong answer;
the direct route does not lean on ARPACK internals. The mode gate is
mutation-tested: a test forbids the range restart under SA and fails if the
branch is removed.

+++ mdopt/mps/explicit.py
@@ -510,7 +510,7 @@ class ExplicitMPS:
     def compress_bond(
         self,
         bond: int,
-        chi_max: int = int(1e4),
+        chi_max: float = int(1e4),
         cut: float = float(1e-17),
         renormalise: bool = False,
         return_truncation_error: bool = False,
@@ -523,7 +523,7 @@ class ExplicitMPS:
         ----------
         bond : int
             The index of the bond to compress.
-        chi_max : int
+        chi_max : float
             The maximum bond dimension to keep.
         cut : float
             Singular values smaller than this will be discarded.
@@ -593,7 +593,7 @@ class ExplicitMPS:
 
     def compress(
         self,
-        chi_max: int = int(1e4),
+        chi_max: float = int(1e4),
         cut: float = float(1e-17),
         renormalise: bool = False,
         return_truncation_errors: bool = False,
@@ -603,7 +603,7 @@ class ExplicitMPS:
 
         Parameters
         ----------
-        chi_max : int
+        chi_max : float
             The maximum bond dimension to keep.
         cut : float
             Singular values smaller than this will be discarded.
